deployment /requirements.txt
streamlit
pandas 
numpy
matplotlib
seaborn 
Pillow
scikit-learn
imblearn
feature_engine
